Move tiger specific functions to separate file
Constructor was already too complex. Even Sonar started to complain.

JIRA: TNT-879
